+++ allst-springboot-exercise/src/main/java/com/allst/exercise/service/CustomersService.java
@@ -9,5 +9,5 @@ import com.allst.exercise.model.Customers;
 public interface CustomersService {
     String save(Customers customers);
 
-    String queryOne(Integer id);
+    String queryOne(Long id);
 }
